iter9 - refactored - moved all logic from main to app package
