+++ cmd/agent/main.go
@@ -1,49 +1,7 @@
 package main
 
-import (
-	"github.com/amangeldi0/metric-tracker/internal/agent/config"
-	"github.com/amangeldi0/metric-tracker/internal/agent/metrics"
-	metricsupdater "github.com/amangeldi0/metric-tracker/internal/agent/metrics_updater"
-	"github.com/go-resty/resty/v2"
-	"go.uber.org/zap"
-	"time"
-)
+import "github.com/amangeldi0/metric-tracker/internal/agent/app"
 
 func main() {
-	logger, err := zap.NewDevelopment()
-	if err != nil {
-		panic(err)
-	}
-
-	defer func(logger *zap.Logger) {
-		if err = logger.Sync(); err != nil {
-			panic(err)
-		}
-	}(logger)
-
-	sugarLogger := logger.Sugar()
-
-	config.Load()
-	if err = config.Parse(); err != nil {
-		sugarLogger.Panicf("Failed loading config: %s", err)
-	}
-
-	client := resty.New()
-	store := metrics.NewRuntimeMetrics()
-
-	go func() {
-		for {
-			time.Sleep(time.Second * time.Duration(config.Config.PollInterval))
-
-			if err = store.Update(); err != nil {
-				sugarLogger.Panicf("Failed to update metrics: %s", err)
-			}
-		}
-	}()
-
-	updater := metricsupdater.New(client, store, sugarLogger)
-	for {
-		time.Sleep(time.Second * time.Duration(config.Config.ReportInterval))
-		updater.UpdateMetrics()
-	}
+	app.Run()
 }
